fix(auth): 认证链路限流原子化与注册/会话并发修复
变更点：
- 限流 Get+Incr 两步操作合并为 Lua 脚本原子 Allow，修复并发计数竞态
- 限流计数全链路 int8→int64，修复计数溢出导致限流失效
- 注册区分「用户不存在」与 DB 故障，避免故障被误判为邮箱可用
- 创建用户识别 MySQL 唯一索引冲突(1062)为邮箱已存在，兜底并发注册竞态
- 登录按错误类型分级记录服务端日志，保留统一文案防用户枚举
- 登出删除会话限定 user_id+token 双条件，避免误删其他用户会话
- 会话表 user_id 改唯一索引并改用 OnConflict upsert，消除并发孤儿会话行
- Authorization 头解析兼容 Bearer/bearer 大小写与多余空白
- /auth/signin 独立限流桶，避免 checkin 等操作挤占登录额度
- 登录/注册接口邮箱增加格式校验

其他：
- 更新 go-sql-driver/mysql 从 indirect 依赖转为直接依赖
- 新增 docs/fix-reports/auth-flow-fixes.md 认证链路修复报告

+++ application/auth/app.go
@@ -27,7 +27,7 @@ type AuthApp interface {
 	Validate(ctx context.Context, token string) (domaintypes.UserID, error)
 
 	// 处理用户限流
-	RateLimit(ctx context.Context, clientIP string, limit int8) error
+	RateLimit(ctx context.Context, clientIP string, limit int64) error
 }
 
 // authAppImpl 认证应用实现
